feat: Add EnvLogLevel function
This commit adds a new EnvLogLevel() function that reads the
LOG_LEVEL environment variable and returns a corresponding
log level.

LOG_LEVEL can now be set to:

  -1, trace -> log.TRACE
  0, debug -> log.DEBUG
  1, info -> log.INFO
  2, warning, warn -> log.WARNING
  3, error -> log.ERROR
  4, fatal -> log.FATAL

LOG_LEVEL is case insensitive.

If LOG_LEVEL is unset, it will default to log.WARNING.

This allows the log level to be configured via an environment
variable.